+++ compose/lib.h
@@ -34,13 +34,15 @@
 #ifndef MUTT_COMPOSE_LIB_H
 #define MUTT_COMPOSE_LIB_H
 
+#include <stdint.h>
+
 struct Buffer;
 struct Email;
 
 /* flags for mutt_compose_menu() */
 #define MUTT_COMPOSE_NOFREEHEADER (1 << 0)
 
-int mutt_compose_menu(struct Email *e, struct Buffer *fcc, struct Email *e_cur, int flags, struct ConfigSubset *sub);
+int mutt_compose_menu(struct Email *e, struct Buffer *fcc, struct Email *e_cur, uint8_t flags, struct ConfigSubset *sub);
 
 bool config_init_compose(struct ConfigSet *);
